requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.11.1
image==1.5.5
kombu==3.0.35
olefile==0.44
Pillow==4.1.1
pytz==2016.6.1
qrcode==5.3
redis==2.10.5
six==1.10.0
